requirements/requrements.txt
Django==1.11.3
django-cors-headers==2.1.0
django-debug-toolbar==1.8
django-extra-fields==0.9
django-filter==1.0.4
djangorestframework==3.6.3
gunicorn==19.7.1
PyMySQL==0.7.11
requests==2.17.3
django-redis==4.8.0
redis==2.10.6
pycrypto==2.6.1
Pillow==5.0.0
xlwt==1.3.0
incremental==17.5.0
pymongo==3.6.1
elasticsearch==6.2.0
celery==4.2.1
#django-celery-results==1.0.1
sqlalchemy==1.2.0
#pip install flower
qrcode==5.3
itsdangerous==0.24